issue #161: optimized value extraction (binary operations are more efficient to convert BCD values)
